requirements.txt
blinker==1.4
bottle==0.12.19
bottle-websocket==0.2.9
cffi==1.15.0
click==8.0.2
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Moment==1.0.2
Flask-WTF==0.15.1
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.3.0
pyparsing==3.0.4
typing-extensions==3.10.0.2
Werkzeug==2.0.2
whichcraft==0.6.1
WTForms==2.3.3
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
